+++ packages/manager/cypress/support/ui/events.ts
@@ -4,5 +4,8 @@
  * @param message - Message that should be displayed on toast notification.
  */
 export const assertToast = (message: string) => {
+  cy.log(
+    'Deprecation warning: Use `ui.toast.assertMessage(...)` instead of `assertToast(...)`.'
+  );
   cy.contains('[data-qa-toast]', message).should('be.visible');
 };