+++ packages/manager/cypress/support/ui/index.ts
@@ -1,15 +1,23 @@
+import * as actionMenu from './action-menu';
 import * as buttons from './buttons';
 import * as dialog from './dialog';
 import * as drawer from './drawer';
 import * as entityHeader from './entity-header';
+import * as heading from './heading';
+import * as nav from './nav';
 import * as tabList from './tab-list';
+import * as toast from './toast';
 import * as toggle from './toggle';
 
 export const ui = {
+  ...actionMenu,
   ...buttons,
   ...dialog,
   ...drawer,
   ...entityHeader,
+  ...heading,
+  ...nav,
+  ...toast,
   ...tabList,
   ...toggle,
 };