+++ packages/manager/src/components/SupportError.tsx
@@ -9,14 +9,28 @@ interface Props {
 
 export const SupportError: React.FC<Props> = (props) => {
   const { errors } = props;
-  const [reason] = errors[0].reason.split(/open a support ticket\./i);
+  const errorMsg = errors[0].reason.split(/(open a support ticket)/i);
 
   const classes = useStyles();
 
   return (
     <Typography className={classes.noticeText}>
-      {reason} {` `}
-      <SupportLink text="open a support ticket" />.
+      {errorMsg.map((substring: string) => {
+        const openTicket = substring.match(/open a support ticket/i);
+        if (openTicket) {
+          return (
+            <SupportLink
+              text={
+                substring.match(/Open.*/)
+                  ? 'Open a support ticket'
+                  : 'open a support ticket'
+              }
+            />
+          );
+        } else {
+          return substring;
+        }
+      })}
     </Typography>
   );
 };