+++ packages/manager/src/factories/managed.ts
@@ -1,16 +1,32 @@
 import * as Factory from 'factory.ts';
 import {
   DataSeries,
+  ManagedContact,
   ManagedCredential,
   ManagedIssue,
+  ManagedLinodeSetting,
   ManagedServiceMonitor,
+  ManagedSSHPubKey,
+  ManagedSSHSetting,
   ManagedStats,
 } from '@linode/api-v4/lib/managed/types';
+
+export const contactFactory = Factory.Sync.makeFactory<ManagedContact>({
+  email: Factory.each((i) => `john.doe.${i}@example.com`),
+  group: 'on-call',
+  id: Factory.each((i) => i + 100),
+  name: Factory.each((i) => `John Doe ${i}`),
+  phone: {
+    primary: '555-555-5555',
+    secondary: null,
+  },
+  updated: '2019-08-01T20:29:14',
+});
 
 export const credentialFactory = Factory.Sync.makeFactory<ManagedCredential>({
   id: Factory.each((i) => i),
   last_decrypted: '2019-07-01',
-  label: 'credential-1',
+  label: Factory.each((i) => `Credential ${i}`),
 });
 
 export const monitorFactory = Factory.Sync.makeFactory<ManagedServiceMonitor>({
@@ -24,7 +40,9 @@ export const monitorFactory = Factory.Sync.makeFactory<ManagedServiceMonitor>({
   service_type: 'url',
   notes: '',
   id: Factory.each((i) => i),
-  credentials: credentialFactory.buildList(3),
+  credentials: credentialFactory
+    .buildList(3)
+    .map((credential) => credential.id),
   address: 'http://www.example.com',
   body: '',
 });
@@ -59,3 +77,28 @@ export const managedIssueFactory = Factory.Sync.makeFactory<ManagedIssue>({
   id: 823,
   services: [],
 });
+
+export const managedSSHSettingFactory = Factory.Sync.makeFactory<ManagedSSHSetting>(
+  {
+    access: true,
+    user: 'root',
+    ip: 'any',
+    port: 22,
+  }
+);
+
+export const managedLinodeSettingFactory = Factory.Sync.makeFactory<ManagedLinodeSetting>(
+  {
+    id: Factory.each((i) => i),
+    label: Factory.each((i) => `Managed Linode ${i}`),
+    group: 'linodes',
+    ssh: managedSSHSettingFactory.build(),
+  }
+);
+
+export const managedSSHPubKeyFactory = Factory.Sync.makeFactory<ManagedSSHPubKey>(
+  {
+    ssh_key:
+      'ssh-rsa MOCKEDSSHKEYMOCKEDSSHKEYMOCKEDSSHKEYMOCKEDSSHKEYMOCKEDSSHKEYMOCKEDSSHKEYMOCKEDSSHKEYMOCKEDSSHKEYMOCKEDSSHKEY managedservices@linode',
+  }
+);