+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2022-11-01] - v1.79.0
+### Added:
+- Set custom UserAgent header for api-v4 when run in node
+
+### Changed:
+- Linode label max characters increased to 64 chars
+- Update Configuration Profile doc link
+- Marketplace app info button can be focused via keyboard
+- Notices that suggest opening a support ticket now include a link to do so
+
+### Fixed:
+- Error when swapping Linode IPs after having already done so
+- Issue preventing more than 100 Marketplace apps from appearing
+- Error when updating billing contact info without a company name in certain circumstances
+
+
 ## [2022-10-17] - v1.78.0
 
 ### Added: