+++ packages/api-v4/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [2022-11-01] - v0.81.0
+### Added:
+- `setUserAgentPrefix` helper function to change API request user agent prefix
+
+### Changed:
+- `credentials` field on `ManagedServiceMonitor` is now an array of numbers
+- API request user agent reflects environment
+
 ## [2022-10-17] - v0.80.0
 
 ### Changed: